Fix problems from old archives.
Case HB-8200

This commit fixes three problems.

1) Some metadata about the mailing list is pickled for use during
archiving a list. When this data was pickled under the previous version
some of the strings could be byte strings. The new version expects all
the data to be strings. We now ensure all the data from the pickle file
are normal strings.

2) Under the old version, when an mbox file is written to for archiving
the encoding of the last email is used for the encoding of the mbox
file. After an upgrade to the new version, some of these mbox files
could be non-utf8 which would cause encoding issues when we try to read
the files. We now catch that error and write the file back to disk as
utf8.

3) When sending an email to the list via sendmail, the messages were
missing a 'content-transfer-encoding' header and caused errors when
Mailman would load these messages. We now ensure messages have this
header and set it to the default value if it is missing.

Changelog: Fix problems from old archives.