Upgrade unpickled Message object for Python 3 compatibility.
Case HB-8201: If a message is held for moderation prior to the Python 3 upgrade,
the resulting pickle file will be missing a required field that email.message
on Python 3 expects. This patch fills in the missing field so that messages held
for moderation prior to the upgrade can still be processed.